+++ crates/camelmailer-core/src/testing.rs
@@ -113,6 +113,7 @@ impl Fixtures {
             name: name.into(),
             verified: true,
             verification_token: token::generate_token(32),
+            check_dmarc: true,
             dkim_private_key: None,
         })
     }
@@ -125,6 +126,7 @@ impl Fixtures {
             name: name.into(),
             verified: false,
             verification_token: token::generate_token(32),
+            check_dmarc: true,
             dkim_private_key: None,
         })
     }